Organizational Changes
Placed scripts in folders. Made new floor prefab, and kept making changes to level 1
