+++ Assets/Scripts/EnemyScripts/EnemyChase.cs
@@ -1,0 +1,25 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EnemyChase : MonoBehaviour
+{
+    public Transform player;
+    public float moveSpeed = 5f;
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        if (player != null)
+        {
+            Vector3 direction = (player.position - transform.position).normalized;
+
+            transform.position += direction * moveSpeed * Time.deltaTime;
+        }
+    }
+}
