+++ Assets/Scripts/EnemyScripts/EnemyCounter.cs
