+++ Assets/Scripts/MenusAndPickups.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 64ab0f77425af4546a159760f2ccd5b2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
