+++ Assets/Scripts/MovingPlatformScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e88c133fe4635134ba57e1c8fe59e875
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
