+++ Assets/Scripts/WaypointPath.cs
@@ -1,0 +1,23 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class WaypointPath : MonoBehaviour
+{
+    public Transform GetWaypoint (int waypointIndex)
+    {
+        return transform.GetChild(waypointIndex);
+    }
+
+    public int GetNextWayPointIndex(int currentWaypointIndex)
+    {
+        int nextWaypointIndex = currentWaypointIndex + 1;
+
+        if (nextWaypointIndex == transform.childCount)
+        {
+            nextWaypointIndex = 0;
+        }
+
+        return nextWaypointIndex;
+    }
+}
