+++ Assets/Scripts/WaypointPath.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 94d8c2012010c014c9c2b09ccf1801db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
